requirements.txt
Flask==3.1.2
python-dotenv==1.1.1
torch==2.8.0
torchvision==0.23.0
safetensors==0.6.2
packaging==25.0
diffusers==0.35.1
transformers==4.56.2
accelerate==1.10.1
tzlocal==5.3.1
peft==0.17.1
torchao==0.13.0
gguf==0.17.1
ftfy==6.3.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
opencv-python==4.12.0.88